// src/interfaces/fantlab/IFantlabWork.ts
export interface IFantlabWork {
    autor1_id: number;
    autor1_rusname: string;
    autor2_id: number;
    autor2_rusname: string;
    autor3_id: number;
    autor3_rusname: string;
    autor4_id: number;
    autor4_rusname: string;
    autor5_id: number;
    autor5_rusname: string;
    name_show_im: string;
    rusname: string;
    work_id: string;
    work_type_id: number;
    pic_edition_id_auto: string;
}