// src/interfaces/fantlab/IFantlabEdition.ts
export interface IFantlabEdition {
  autors: string;
  comment: string;
  compilers: string;
  correct: number;
  doc: number;
  edition_id: number;
  isbn1: string;
  isbn2: string;
  name: string;
  notes: string;
  plan_date: string;
  publisher: string;
  series: string;
  weight: number;
  year: string;
}